Move to gallery.js
